# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Webservice SI-ASN BKN Gateway API",
    "version": "1.0.0",
    "description": "### Dokumentasi Gateway Integrasi Web Service SI-ASN BKN (SIASNAPI-SIMPEG)\nGateway ini bertindak sebagai **Reverse Proxy** yang mengurus autentikasi Dual-Token BKN (WSO2 OAuth2 & SSO) secara transparan.\n\n**Panduan Penggunaan untuk Pengembang Aplikasi Daerah:**\n1. Semua request dikirim ke Gateway ini tanpa perlu pusing menyisipkan token BKN.\n2. Gateway otomatis menambahkan header `Authorization: Bearer <wso2_token>` dan `Auth: Bearer <sso_token>`.\n3. URL dasar gateway dapat dipanggil menggunakan prefix `/apisiasn/1.0` atau `/api`.\n4. Untuk unduh berkas gunakan endpoint `/download-dok?filePath=...` yang akan langsung mengalirkan file PDF dari BKN."
  },
  "servers": [
    {
      "url": "/apisiasn/1.0",
      "description": "Default Proxy Gateway (/apisiasn/1.0)"
    },
    {
      "url": "/api",
      "description": "Shorthand Proxy Gateway (/api)"
    }
  ],
...
  "paths": {
    "/pns/data-utama/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Data Utama PNS Lengkap",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/data-anak/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Data Anak PNS by NIP",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/data-pasangan/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Data Pasangan PNS by NIP",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-jabatan/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Jabatan PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-golongan/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Golongan / Pangkat PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-pendidikan/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Pendidikan PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-diklat/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Diklat Struktural PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-kursus/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Kursus PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-hukdis/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Hukuman Disiplin PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-skp/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat SKP PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-skp22/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat SKP 2022 PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-kinerjaperiodik/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Kinerja Periodik PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-angkakredit/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Angka Kredit PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-cltn/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Cuti di Luar Tanggungan Negara (CLTN)",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-kontrak/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Kontrak PPPK",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-potensi/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Potensi PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-kompetensi/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Kompetensi PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-dp3/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat DP3 PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-sertifikasi/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Sertifikasi PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-masakerja/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Masa Kerja PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-kgb/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Kenaikan Gaji Berkala (KGB) PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-pemberhentian/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Pemberhentian PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-tubel/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Tugas Belajar (Tubel) PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-penghargaan/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Penghargaan PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-pindahinstansi/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat Pindah Instansi PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-pnsunor/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat PNS Unor",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/rw-pwk/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Riwayat PWK PNS",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/nilaiipasn/{nipBaru}": {
      "get": {
        "tags": [
          "Data Utama & Riwayat PNS"
        ],
        "summary": "Get Nilai IP ASN",
        "parameters": [
          {
            "name": "nipBaru",
            "in": "path",
            "required": true,
            "description": "NIP Pegawai (18 digit)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
...
    "/pns/list-pensiun-instansibynip": {
      "get": {
        "tags": [
          "Pemberhentian & Pensiun"
        ],
        "summary": "Get List Usulan Pensiun Berdasarkan NIP",
        "parameters": [
          {
            "name": "nip",
            "in": "query",
            "required": true,
            "description": "NIP Pegawai",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/list-pensiun-instansi": {
      "get": {
        "tags": [
          "Pemberhentian & Pensiun"
        ],
        "summary": "Get List Usulan Pensiun Instansi by Rentang Tanggal",
        "parameters": [
          {
            "name": "tglAwal",
            "in": "query",
            "required": true,
            "description": "dd-mm-yyyy",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tglAkhir",
            "in": "query",
            "required": true,
            "description": "dd-mm-yyyy",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/kp-instansi-byId": {
      "get": {
        "tags": [
          "Kenaikan Pangkat (KP)"
        ],
        "summary": "Get Status Usulan KP Instansi by ID Usulan",
        "parameters": [
          {
            "name": "idUsulan",
            "in": "query",
            "required": true,
            "description": "ID Usulan KP BKN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
    "/pns/list-kp-instansi": {
      "get": {
        "tags": [
          "Kenaikan Pangkat (KP)"
        ],
        "summary": "Get List Usulan KP Instansi by Periode",
        "parameters": [
          {
            "name": "periode",
            "in": "query",
            "required": true,
            "description": "yyyy-mm-dd",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "description": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "description": "offset",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Berhasil menerima respon dari BKN SIASN"
          },
          "400": {
            "description": "Parameter request tidak valid"
          },
          "404": {
            "description": "Data tidak ditemukan di BKN"
          },
          "500": {
            "description": "Kesalahan internal server BKN"
          }
        }
      }
    },
...
  }
}